admin: split admin commands in a plugins.admins package
