+++ eduu/plugins/__init__.py
@@ -1,3 +1,3 @@
-"""EduuRobot Plugins!"""
+"""EduuRobot plugins package."""
 # SPDX-License-Identifier: MIT
 # Copyright (c) 2018-2022 Amano Team
