+++ eduu/plugins/admins/__init__.py
@@ -1,5 +1,3 @@
-"""EduuRobot core package."""
+"""EduuRobot plugins.admins package."""
 # SPDX-License-Identifier: MIT
 # Copyright (c) 2018-2022 Amano Team
-
-__version__ = "2.0.0-beta"
