+++ eduu/utils/__init__.py
@@ -15,7 +15,9 @@ from .utils import (
     del_restarted,
     get_emoji_regex,
     get_format_keys,
+    get_reason_text,
     get_restarted,
+    get_target_user,
     http,
     pretty_size,
     remove_escapes,
@@ -39,7 +41,9 @@ __all__: List[str] = [
     "del_restarted",
     "get_emoji_regex",
     "get_format_keys",
+    "get_reason_text",
     "get_restarted",
+    "get_target_user",
     "http",
     "pretty_size",
     "remove_escapes",
